Fixed Status updating screen

+++ app/src/main/java/com/elfennani/aniwatch/domain/models/StatusDetails.kt
@@ -4,7 +4,7 @@ import com.elfennani.aniwatch.domain.models.enums.ShowStatus
 
 data class StatusDetails(
     val status: ShowStatus?,
-    val score: Int,
+    val score: Double,
     val progress: Int,
     val favorite: Boolean,
     val startedAt: StatusDate?,
